Bash completion: Add cache to speed up completion
The command

  gulp --tasks-simple

is executed for completion on tab press. It may take some time to
execute, resulting in an annoying lag.

To work around this, bash completion now uses a cache file that contains
the latest result of the above command. The cache invalidates as soon as
the gulpfile last modified time changes.